Update and delete functionality. Also username validation. Project completed.

+++ routes/ideas.js
@@ -58,22 +58,22 @@ router.post('/', async (req, res) => {
 
 router.put('/:id', async (req, res) => {
   try {
-    const updatedIdea = await Idea.findByIdAndUpdate(
-      req.params.id,
-      {
-        $set: {
-          text: req.body.text,
-          tag: req.body.tag
-        }
-      },
-      {
-        new: true
-      }
-    )
-    res.json({
-      success: true,
-      data: updatedIdea
-    })  
+    const idea = await Idea.findById(req.params.id)
+
+    if (idea.username === req.body.username) {
+      const updatedIdea = await Idea.findByIdAndUpdate(
+        req.params.id,
+        {
+          $set: {
+            text: req.body.text,
+            tag: req.body.tag
+          }
+        },
+        { new: true }
+      )
+      return res.json({ success: true, data: updatedIdea })  
+    }
+    res.status(403).json({ success: false, error: 'You are not authorised to update this record.' })
   } catch (error) {
     console.log(error)
     res.status(500).json({
@@ -85,17 +85,17 @@ router.put('/:id', async (req, res) => {
 
 router.delete('/:id', async (req, res) => {
   try {
-    await Idea.findByIdAndDelete(req.params.id)
-    res.json({
-      success: true,
-      data: {}
-    })
+    const idea = await Idea.findById(req.params.id)
+
+    if (idea.username === req.body.username) {
+      await Idea.findByIdAndDelete(req.params.id)
+      return res.json({ success: true, data: {} })
+    }
+
+    res.status(403).json({ success: false, error: 'You are not authorised to delete this record.' })
   } catch (error) {
     console.log(error)
-    res.status(500).json({
-      success: false,
-      error: 'Something went wrong.'
-    })
+    res.status(500).json({ success: false, error: 'Something went wrong.' })
   }
 })
